+++ 404.php
@@ -9,10 +9,11 @@ get_header(); ?>
 
 	<div id="primary" class="content-area container content">
 		<main id="main" class="site-main row" role="main">
+			<?php get_template_part( 'template-parts/sidebar', 'left' ); ?>
 			<?php
 				$col = 12;
-				if ( is_active_sidebar( 'sidebar-1' ) ) $col -= 3;
-				if ( is_active_sidebar( 'content' ) ) $col -= 3;
+				if ( is_active_sidebar( 'left-sidebar' ) ) $col -= 3;
+				if ( is_active_sidebar( 'right-sidebar' ) ) $col -= 3;
 			?>
 			<section class="error-404 not-found col-md-<?php echo $col; ?>">
 				<header class="page-header">
@@ -54,11 +55,7 @@ get_header(); ?>
 				</div><!-- .page-content -->
 			</section><!-- .error-404 -->
 
-			<?php if ( is_active_sidebar( 'content' ) ) : ?>
-			<section class="col-md-3">
-				<?php get_sidebar( 'content' ); ?>
-			</section><!-- .col-md-3 -->
-			<?php endif; ?>
+			<?php get_template_part( 'template-parts/sidebar', 'right' ); ?>
 
 		</main><!-- #main -->
 	</div><!-- #primary -->