+++ footer.php
@@ -1,9 +1,9 @@
 		</div><!-- .main .shadow -->
 			<div class="container bottom-blocks"><!-- Example row of columns -->
-				<?php if ( is_active_sidebar( 'footer-widget' ) ) : ?>
-				<div id="footer-widget" class="footer-widget widget-area row fill" role="complementary">
-					<?php dynamic_sidebar( 'footer-widget' ); ?>
-				</div><!-- #footer-widget -->
+				<?php if ( is_active_sidebar( 'bottom-widget' ) ) : ?>
+				<div id="bottom-widget" class="bottom-widget widget-area row fill" role="complementary">
+					<?php dynamic_sidebar( 'bottom-widget' ); ?>
+				</div><!-- #bottom-widget -->
 				<?php endif; ?>
 			</div>
 		</div><!-- /container -->