+++ header.php
@@ -51,6 +51,10 @@
 					 	</div>
 						<nav id="primary-navigation" class="site-navigation primary-navigation" role="navigation">
 							<?php
+								if ( 0 == get_option( 'dswoddil_cache_menu' ) ) {
+									delete_transient( 'dswoddil_primary_menu_data' );
+								}
+
 								if ( false === ( $dswoddil_menu_data = get_transient( 'dswoddil_primary_menu_data' ) ) ) {
 									$dswoddil_menu_data = wp_nav_menu( array(
 										'echo'				=> false,
@@ -64,8 +68,17 @@
 										'fallback_cb'       => 'wp_bootstrap_navwalker::fallback',
 										'walker'            => new wp_bootstrap_navwalker())
 									);
-									set_transient( 'dswoddil_primary_menu_data', $dswoddil_menu_data, 10 * MINUTE_IN_SECONDS );
+									if ( 0 < get_option( 'dswoddil_cache_menu' ) ) {
+										set_transient(
+											'dswoddil_primary_menu_data',
+											$dswoddil_menu_data,
+											get_option( 'dswoddil_cache_menu' ) * MINUTE_IN_SECONDS
+										);
+									} else {
+										delete_transient( 'dswoddil_primary_menu_data' );
+									}
 								}
+
 								echo $dswoddil_menu_data;
 							?>
 						</nav>
@@ -73,3 +86,9 @@
 				</div>
 				<?php } ?>
 			</div>
+
+			<?php if ( is_active_sidebar( 'top-widget' ) ) : ?>
+			<div id="top-widget" class="top-widget widget-area container optional" role="complementary">
+				<?php dynamic_sidebar( 'top-widget' ); ?>
+			</div><!-- #content-widget -->
+			<?php endif; ?>