+++ index.php
@@ -16,12 +16,6 @@
 
 get_header(); ?>
 
-	<?php if ( is_active_sidebar( 'above-content-widget' ) ) : ?>
-		<div id="above-content-widget" class="above-content-widget widget-area container optional" role="complementary">
-			<?php dynamic_sidebar( 'above-content-widget' ); ?>
-		</div><!-- #content-widget -->
-	<?php endif; ?>
-
 	<?php
 		if ( is_front_page() && dswoddil_has_featured_posts() ) {
 			// Include the featured content template.
@@ -31,17 +25,11 @@ get_header(); ?>
 
 	<div id="primary" class="content-area container content">
 		<main id="main" class="site-main row" role="main">
-		<?php if ( is_active_sidebar( 'sidebar-1' ) ) : ?>
-			<section class="col-md-3">
-				<div id="primary-sidebar" class="primary-sidebar widget-area" role="complementary">
-					<?php dynamic_sidebar( 'sidebar-1' ); ?>
-				</div><!-- #primary-sidebar -->
-			</section>
-		<?php endif; ?>
+		<?php get_template_part( 'template-parts/sidebar', 'left' ); ?>
 		<?php
 			$col = 12;
-			if ( is_active_sidebar( 'sidebar-1' ) ) $col -= 3;
-			if ( is_active_sidebar( 'content' ) ) $col -= 3;
+			if ( is_active_sidebar( 'left-sidebar' ) ) $col -= 3;
+			if ( is_active_sidebar( 'right-sidebar' ) ) $col -= 3;
 		?>
 			<section class="col-md-<?php echo $col; ?>">
 				<?php
@@ -66,11 +54,7 @@ get_header(); ?>
 
 				?>
 			</section><!-- .col-md-9 -->
-			<?php if ( is_active_sidebar( 'content' ) ) : ?>
-			<section class="col-md-3">
-				<?php get_sidebar( 'content' ); ?>
-			</section><!-- .col-md-3 -->
-			<?php endif; ?>
+			<?php get_template_part( 'template-parts/sidebar', 'right' ); ?>
 		</main><!-- #main .row -->
 	</div><!-- #primary .content -->